working on calculation
